sort 초안

+++ src/error_check.c
@@ -13,7 +13,7 @@
 #include "../includes/push_swap.h"
 
 // 중복, int의 최대 최소 체크
-int		argv_check(int arr[], int num)
+int		argv_check(long long arr[], int num)
 {
 	int idx;
